requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pillow>=10.3.0
numpy>=1.26.0
scipy>=1.13.0
pyyaml>=6.0.1
jinja2>=3.1.3
python-multipart>=0.0.9
